Fetch active streams, 100 at a time (#2)
Use the twitch streams api (https://dev.twitch.tv/docs/v5/reference/streams) to fetch all the active streams. There are usually around 20k streams live at a time, but you can only fetch 100 at a time. There is also a limit of 1 request per second, so this required creating a rate limiter that dispatches queued api calls. Once an api call is completed, the resulting channel is stored. It continually cycles through all the streams, and if a stored stream does not show up in a cycle, it is removed.